add router, controller and refactor server.js... controller now holds majority of logic communicating with db and client
